# lesson-booking/openapi.yaml
openapi: 3.1.0
info:
  title: Lesson Booking API
  version: 0.1.0
  description: |
    レッスン予約サービス。複数 Google カレンダー連携で空き枠を出し、40 日先まで予約できる。
    レッスン開始まで 14 日未満のキャンセル・変更はメッセージ+対応方法(3 択)を添えた変更要求となり、主催者の承認で確定する。
    詳細は docs/spec.md。
servers:
  - url: http://localhost:8787
security:
  - bearerAuth: []
  - devHeader: []
...
paths:
...
  /hosts:
    post:
      tags: [host]
      summary: ログイン中ユーザーを主催者として登録
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [displayName]
              properties:
                displayName: { type: string }
                slug: { type: string, pattern: '^[a-z0-9-]{3,32}$', description: 公開予約ページの URL 名。省略時はランダム }
                bio: { type: string }
                timezone: { type: string, default: Asia/Tokyo }
                lessonMinutes: { type: integer, default: 60 }
                minLeadMinutes: { type: integer, default: 60 }
                rescheduleRangeDays: { type: integer, minimum: 1, maximum: 30, default: 7 }
                lateChangeThresholdDays: { type: integer, minimum: 0, maximum: 60, default: 14 }
                bookingHorizonDays: { type: integer, minimum: 1, maximum: 60, default: 40 }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Host' } } } }
        '409': { $ref: '#/components/responses/Error' }
  /hosts/{hostId}:
    parameters: [{ $ref: '#/components/parameters/hostId' }]
    patch:
      tags: [host]
      summary: 主催者設定の変更
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName: { type: string }
                slug: { type: string, pattern: '^[a-z0-9-]{3,32}$' }
                bio: { type: string }
                cancellationFeeAmount: { type: [integer, 'null'], minimum: 50 }
                timezone: { type: string }
                lessonMinutes: { type: integer }
                minLeadMinutes: { type: integer }
                rescheduleRangeDays: { type: integer, minimum: 1, maximum: 30, description: 振替期間(元のレッスン日から前後の日数) }
                lateChangeThresholdDays: { type: integer, minimum: 0, maximum: 60, description: 開始の何日前から承認制にするか }
                bookingHorizonDays: { type: integer, minimum: 1, maximum: 60, description: 何日先まで予約を受け付けるか }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Host' } } } }
...
  /hosts/{hostId}/calendars:
    parameters: [{ $ref: '#/components/parameters/hostId' }]
    get:
      tags: [host]
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/HostCalendar' } } } } }
    post:
      tags: [host]
      summary: 連携カレンダー追加(write_target は 1 件のみ)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [calendarId, role]
              properties:
                calendarId: { type: string, example: primary }
                label: { type: string }
                role: { type: string, enum: [busy_source, write_target] }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/HostCalendar' } } } }
        '400': { $ref: '#/components/responses/Error' }
  /hosts/{hostId}/calendars/{id}:
    parameters: [{ $ref: '#/components/parameters/hostId' }, { $ref: '#/components/parameters/id' }]
    delete:
      tags: [host]
      responses:
        '204': { description: No Content }
  /hosts/{hostId}/availability-windows:
    parameters: [{ $ref: '#/components/parameters/hostId' }]
    get:
      tags: [host]
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/AvailabilityWindow' } } } } }
    post:
      tags: [host]
      summary: 営業時間枠追加
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [weekday, startTime, endTime]
              properties:
                weekday: { type: integer, minimum: 0, maximum: 6, description: 0=日曜 }
                startTime: { type: string, pattern: '^\d{2}:\d{2}$', example: '10:00' }
                endTime: { type: string, pattern: '^\d{2}:\d{2}$', example: '18:00' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/AvailabilityWindow' } } } }
        '400': { $ref: '#/components/responses/Error' }
  /hosts/{hostId}/availability-windows/{id}:
    parameters: [{ $ref: '#/components/parameters/hostId' }, { $ref: '#/components/parameters/id' }]
    delete:
      tags: [host]
      responses:
        '204': { description: No Content }
  /hosts/{hostId}/bookings:
    parameters: [{ $ref: '#/components/parameters/hostId' }]
    get:
      tags: [host]
      summary: 主催者の予約(生徒情報付き)。過去分は 60 日前まで、キャンセルフィー未払いは期間に関係なく含む
      parameters: [{ $ref: '#/components/parameters/since' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - { $ref: '#/components/schemas/Booking' }
                    - type: object
                      properties:
                        student: { $ref: '#/components/schemas/StudentSummary' }
  /hosts/{hostId}/bookings/{id}/calendar-sync:
    parameters: [{ $ref: '#/components/parameters/hostId' }, { $ref: '#/components/parameters/id' }]
    post:
      tags: [host]
      summary: Google カレンダーへの反映(作成・日時変更・削除)をすぐに再試行する。失敗が続けば calendarSyncError が残ったまま返る
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Booking' } } } }
        '403': { $ref: '#/components/responses/Error' }
  /hosts/{hostId}/bookings/{id}/fee-paid:
    parameters: [{ $ref: '#/components/parameters/hostId' }, { $ref: '#/components/parameters/id' }]
    post:
      tags: [host]
      summary: キャンセルフィーの入金確認
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Booking' } } } }
        '409': { $ref: '#/components/responses/Error' }
  /hosts/{hostId}/change-requests/count:
    parameters: [{ $ref: '#/components/parameters/hostId' }]
    get:
      tags: [host]
      summary: 承認待ちの件数(タブのバッジ用。空き確認はしない)
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { pending: { type: integer } } } } } }
  /hosts/{hostId}/change-requests:
    parameters: [{ $ref: '#/components/parameters/hostId' }]
    get:
      tags: [host]
      summary: 変更要求一覧(既定は pending)
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [pending, approved, rejected, all], default: pending } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - { $ref: '#/components/schemas/ChangeRequest' }
                    - type: object
                      properties:
                        optionLabel: { type: string }
                        feeMethodLabel: { type: [string, 'null'] }
                        candidates:
                          type: array
                          description: 承認待ちの振替の希望日時と、現在空いているか
                          items:
                            type: object
                            properties:
                              startAt: { type: string, format: date-time }
                              available: { type: boolean }
                        booking: { $ref: '#/components/schemas/Booking' }
                        student: { $ref: '#/components/schemas/StudentSummary' }
  /hosts/{hostId}/change-requests/{id}/decision:
    parameters: [{ $ref: '#/components/parameters/hostId' }, { $ref: '#/components/parameters/id' }]
    post:
      tags: [host]
      summary: 変更要求の承認・却下
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [approve, reject] }
                note: { type: string }
                startAt: { type: string, format: date-time, description: 振替の承認で、希望日時の中から選んだ振替先(候補が 1 つなら省略可) }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  request: { $ref: '#/components/schemas/ChangeRequest' }
                  booking: { $ref: '#/components/schemas/Booking' }
        '409': { $ref: '#/components/responses/Error' }
...
  /hosts/{hostId}/bookings/{id}/fee-method:
    parameters: [{ $ref: '#/components/parameters/hostId' }, { $ref: '#/components/parameters/id' }]
    post:
      tags: [host]
      summary: 未払いキャンセルフィーの支払い方法を変更(生徒に案内メール)
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [method], properties: { method: { $ref: '#/components/schemas/FeeMethod' } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Booking' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /hosts/{hostId}/bookings/{id}/cancel:
    parameters: [{ $ref: '#/components/parameters/hostId' }, { $ref: '#/components/parameters/id' }]
    post:
      tags: [host]
      summary: 主催者による取り消し(休講)。生徒にメール通知
      description: 承認待ちの変更要求がある予約は 409 change_request_pending(先に承認/却下する)。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, description: 生徒へのメッセージ }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Booking' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /hosts/{hostId}/billing:
    parameters: [{ $ref: '#/components/parameters/hostId' }]
    get:
      tags: [host]
      summary: プラン・上限・今月の利用量・公開 URL
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  plan: { type: string, enum: [free, pro] }
                  effectivePlan: { type: string, enum: [free, pro] }
                  planLabel: { type: string }
                  subscriptionStatus: { type: string, enum: [none, active, past_due, canceled] }
                  viaOrganization: { type: boolean, description: 教室プランによってプロ相当になっているか }
                  limits:
                    type: object
                    properties:
                      maxCalendars: { type: [integer, 'null'] }
                      maxBookingsPerMonth: { type: [integer, 'null'] }
                      calendarWrite: { type: boolean }
                  usage:
                    type: object
                    properties:
                      bookingsThisMonth: { type: integer }
                      calendars: { type: integer }
                  publicUrl: { type: string }
  /hosts/{hostId}/billing/checkout:
    parameters: [{ $ref: '#/components/parameters/hostId' }]
    post:
      tags: [host]
      summary: プロプランの決済ページ URL(Stripe Checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [successUrl, cancelUrl]
              properties:
                successUrl: { type: string, format: uri }
                cancelUrl: { type: string, format: uri }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { url: { type: string } } } } } }
  /hosts/{hostId}/billing/portal:
    parameters: [{ $ref: '#/components/parameters/hostId' }]
    post:
      tags: [host]
      summary: 支払い方法・解約の管理ページ URL(Stripe Customer Portal)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [returnUrl]
              properties:
                returnUrl: { type: string, format: uri }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { url: { type: string } } } } } }
        '409': { $ref: '#/components/responses/Error' }
...
  /hosts/{hostId}/connect:
    parameters: [{ $ref: '#/components/parameters/hostId' }]
    get:
      tags: [host]
      summary: Stripe 連携状況
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  available: { type: boolean, description: プランで利用可能か }
                  accountId: { type: [string, 'null'] }
                  chargesEnabled: { type: boolean }
                  active: { type: boolean, description: 生徒がオンラインで支払える状態か }
                  cancellationFeeAmount: { type: [integer, 'null'] }
    delete:
      tags: [host]
      summary: Stripe 連携を外す
      responses:
        '204': { description: No Content }
  /hosts/{hostId}/connect/onboarding:
    parameters: [{ $ref: '#/components/parameters/hostId' }]
    post:
      tags: [host]
      summary: Stripe アカウント作成・オンボーディング URL(プロのみ)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshUrl, returnUrl]
              properties:
                refreshUrl: { type: string, format: uri }
                returnUrl: { type: string, format: uri }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { url: { type: string } } } } } }
        '402': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT, description: Supabase Auth の access_token }
    devHeader: { type: apiKey, in: header, name: x-dev-user-email, description: AUTH_MODE=dev のときのみ }
  parameters:
    hostId: { name: hostId, in: path, required: true, schema: { type: string } }
    id: { name: id, in: path, required: true, schema: { type: string } }
    since: { name: since, in: query, description: 過去分の起点(既定は 60 日前), schema: { type: string, format: date-time } }
  responses:
    Error:
      description: エラー
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code:
                    type: string
                    enum: [validation, outside_booking_window, late_change_requires_request, forbidden, not_found, slot_unavailable, change_request_pending, invalid_state, plan_limit, calendar_error, internal]
                  message: { type: string }
                  details: {}
  schemas:
    FeeMethod:
      type: string
      enum: [card, bank_transfer, in_person]
      description: クレジットカード / 銀行振込 / 次回レッスン時に手渡し
    LateChangeOption:
      type: string
      enum: [request_approval, reschedule_within_two_weeks, pay_cancellation_fee]
      description: 承認を求める / 講師の振替期間内の別日に振替(値の名前は互換のため据え置き)/ キャンセルフィーを支払う
...
    Host:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        displayName: { type: string }
        slug: { type: string }
        bio: { type: string }
        plan: { type: string, enum: [free, pro] }
        subscriptionStatus: { type: string, enum: [none, active, past_due, canceled] }
        cancellationFeeAmount: { type: [integer, 'null'] }
        feeMethods: { type: array, items: { $ref: '#/components/schemas/FeeMethod' } }
        bankTransferInfo: { type: string }
        stripeConnectAccountId: { type: [string, 'null'] }
        connectChargesEnabled: { type: boolean }
        organizationId: { type: [string, 'null'] }
        orgPlanActive: { type: boolean, description: 所属教室の契約が有効か(有効ならプロ相当) }
        timezone: { type: string }
        lessonMinutes: { type: integer }
        minLeadMinutes: { type: integer }
        createdAt: { type: string, format: date-time }
    HostCalendar:
      type: object
      properties:
        id: { type: string }
        hostId: { type: string }
        calendarId: { type: string }
        label: { type: string }
        role: { type: string, enum: [busy_source, write_target] }
    AvailabilityWindow:
      type: object
      properties:
        id: { type: string }
        hostId: { type: string }
        weekday: { type: integer }
        startTime: { type: string }
        endTime: { type: string }
...
    StudentSummary:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        name: { type: string }
    Booking:
      type: object
      properties:
        id: { type: string }
        hostId: { type: string }
        studentId: { type: string }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
        status: { type: string, enum: [confirmed, cancelled] }
        calendarEventId: { type: [string, 'null'] }
        note: { type: [string, 'null'] }
        cancellationFeeStatus: { type: string, enum: [none, pending, paid] }
        cancellationFeeAmount: { type: [integer, 'null'] }
        cancellationFeeMethod: { oneOf: [{ $ref: '#/components/schemas/FeeMethod' }, { type: 'null' }] }
        reminderSentAt: { type: [string, 'null'], format: date-time }
        calendarSyncError: { type: [string, 'null'], description: Google カレンダーへの反映に失敗した理由。null なら反映済みか反映不要 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    ChangeRequest:
      type: object
      properties:
        id: { type: string }
        bookingId: { type: string }
        hostId: { type: string }
        studentId: { type: string }
        kind: { type: string, enum: [cancel, reschedule] }
        option: { $ref: '#/components/schemas/LateChangeOption' }
        message: { type: string }
        proposedStartAts: { type: array, items: { type: string, format: date-time }, description: 振替の希望日時(第1希望から順、最大 3) }
        approvedStartAt: { type: [string, 'null'], format: date-time, description: 講師が承認時に選んだ振替先 }
        feeMethod: { oneOf: [{ $ref: '#/components/schemas/FeeMethod' }, { type: 'null' }] }
        status: { type: string, enum: [pending, approved, rejected] }
        decisionNote: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        decidedAt: { type: [string, 'null'], format: date-time }